+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffMacroParameterReferenceMixin.kt
@@ -1,0 +1,21 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.mixin
+
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroParameterReference
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.impl.HuffNamedElementImpl
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffMacroParameterReferenceImpl
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.base.HuffReference
+import com.intellij.lang.ASTNode
+import com.intellij.psi.PsiElement
+
+abstract class HuffMacroParameterReferenceMixin(node: ASTNode) : HuffNamedElementImpl(node), HuffMacroParameterReference {
+
+  override val referenceNameElement: PsiElement
+    get() = this.identifier
+
+  override val referenceName: String
+    get() = referenceNameElement.text
+
+  override fun getName(): String? = referenceName
+
+  override fun getReference(): HuffReference = HuffMacroParameterReferenceImpl(this as HuffMacroParameterReference)
+}