+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffMacroParameterInvocationMixin.kt
@@ -1,0 +1,21 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.mixin
+
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroParameterInvocation
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.impl.HuffNamedElementImpl
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffMacroParameterInvocationReference
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.base.HuffReference
+import com.intellij.lang.ASTNode
+import com.intellij.psi.PsiElement
+
+abstract class HuffMacroParameterInvocationMixin(node: ASTNode) : HuffNamedElementImpl(node), HuffMacroParameterInvocation {
+
+  override val referenceNameElement: PsiElement
+    get() = this.identifier
+
+  override val referenceName: String
+    get() = referenceNameElement.text
+
+  override fun getName(): String? = referenceName
+
+  override fun getReference(): HuffReference = HuffMacroParameterInvocationReference(this as HuffMacroParameterInvocation)
+}