+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffMacroArgumentReferenceMixin.kt
@@ -1,0 +1,21 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.mixin
+
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroArgumentReference
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.impl.HuffNamedElementImpl
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffMacroArgumentReferenceImpl
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.base.HuffReference
+import com.intellij.lang.ASTNode
+import com.intellij.psi.PsiElement
+
+abstract class HuffMacroArgumentReferenceMixin(node: ASTNode) : HuffNamedElementImpl(node), HuffMacroArgumentReference {
+
+  override val referenceNameElement: PsiElement
+    get() = this.identifier
+
+  override val referenceName: String
+    get() = referenceNameElement.text
+
+  override fun getName(): String? = referenceName
+
+  override fun getReference(): HuffReference = HuffMacroArgumentReferenceImpl(this as HuffMacroArgumentReference)
+}